Refactor disaggregation tests
