+++ test/srt/run_suite.py
@@ -138,6 +138,7 @@ suites = {
         # TestFile("ep/test_moe_ep.py", 181),
         TestFile("test_disaggregation.py", 499),
         TestFile("test_disaggregation_different_tp.py", 155),
+        TestFile("test_disaggregation_pp.py", 100),
         TestFile("test_full_deepseek_v3.py", 333),
     ],
     "per-commit-8-gpu-b200": [
